+++ components/coding/studentPortal/LessonComponent.tsx
@@ -65,8 +65,8 @@ export default function LessonComponent({ data }: LessonComponentProps) {
   if (data.component === "code-snippet") {
     return (
       <div className="mx-4 space-y-4">
-        {data.text && <p>{data.text}</p>}
-        <pre className="p-4 border-2 bg-slate-100 border-black-transparent rounded-xl">
+        {data.text && <div>{data.text}</div>}
+        <pre className="p-4 border-2 bg-slate-100 border-black-transparent text-4xl rounded-xl">
           <Sandpack template="react-ts" files={{ "App.tsx": data.code }} />
         </pre>
       </div>
